roa_validate.c: accept date ranges where the end date and the start date are the same (this means the cert is useless, but it's not necessarily invalid). Still disallow certs with retrograde ranges.
myssl.c: we were using the wrong OID for checking the SIA (per the
spec). Also, the URI for a cert that is signing an EE should point to
the publication point of the signed object, which in this case can be
the EE itself (e.g. rsync://mumble/stuff.roa) and not end in a '/', so
remove that test.

sqhl.h: temporarily change the RSYNC_PREFIX we require from "rsync://"
to "rsync:/" so we can load roas from APNIC (they plan to fix this,
and we should revert this change at that point).