extensions.asn: add some oids that we use in myssl
make_test_roa.c: allow invalid roas to be written out

myssl.c: accept both the dir-form and the ee-object-form of sia in
certs. It's not clear to me which form is correct, and our code
generates the dir-form, APNIC's code generates the ee-object-form (and
they wrote the spec). So accept both for now, negotiate the details
with them.