asids no longer found in the EE cert for a roa, so we need to specify it on the command line in make_test_roa, and pass it in in makeroas.
query.c was changed to move a validation check earlier, to make it
more sensible. no real behavioral modification, though.

sqhl.c has a hack in it (my apologies) to extract the embedded cert in
a roa and add it to the database. It generates a fake filename for the
cert. The fake filename used to end in .roa.cer, but this confused the
code that determined the type of an object (resulting in it thinking
it was both a roa and a cert), so I changed it to make generate a fake
filename that ends in .cer (e.g. R111.cer). I should also go back and
fix the function to fail if it finds a confused filename like this
(rather than misbehave).