Dave and I fixed a bug where getuintscm() outputs an unitialized value if there are no rows in the table.
In the process, we discovered that if we ever have enough certificates
that our id's go beyond (2^31 - 1), all sorts of bad things happen.
However, according to Mark, the limitations on the local id are
purely a theoretical problem because the time horizon for it becoming
a practical problem is huge.  Moreover, the system was never intended
to be 64-bit ready.