This is the temporary fix that Andrew mentioned in his e-mails.
For now, this code calls a python script which takes care of our
ipv4 and ipv6 parsing. It slows down the creation process a bit, 
but it is worth it to make it work for now. Also, this is only for 
testbed generation. 

Charlie and I will work on fixing the actual C code.