Revert master's configure.ac to pass 'make test'
As of this commit, the ODBC invalid cursor state issues have not yet
been fixed.  However, for the purpose of having a stable master, I am
reverting back to Brent Kohler's hack: link directly to myodbc and
make sure that -lodbc is at the end of the linker command line.  If
you want to work on the ODBC problem, work on the feature branch
fixForODBC.