Enable stack-protector only if supported.
On roa-pki (FC4), GCC was built with stack-protector support disabled.
Because of this, all subsequent AC_SEARCH_LIBS would fail, not due to
missing libraries, but due to the fact that CFLAGS had been changed to
something illegal.  This got very confusing.

I replaced the GCC test with an explicit compilation test to see
whether the compiler will take the -fstack-protector option.  Now we
turn on the flag only if we are sure the compiler can support it.
Thanks to Dustin Spicuzza for providing a useful example.