Schema changes to prevent duplicate announcements and allow queries to be made in chunks.
NOTE: This changes the meaning of rtr_incremental.serial_num.
It used to be the serial number following the set of announcements and withdrawals,
now it's the serial number preceding the set.
This change makes coding the RTR server easier because Serial Query PDUs request
all announcements and withdrawals *after* a particular serial number.