Change ActionDetail to refuse to show a new Action if the current one is not valid.
Revise AbstractAction to supply the reasons why it is invalid. Implement
invalid reasons in each action.

Revise "Save" and "Run" to refuse to save or run a test with invalid
actions.

Simplify creation of an allocation with only IANA defined. Make both
issuer and subject be the one existing CA (IANA) and depend on the
resulting validity to force the user to define a valid CA for the
subject.

Several fixes to CAEditor and CAChooser to streamline the user
experience and avoid pre-selecting inappropriate values.