add an option to delete all other TAs
I.e. "updateTA.py -d foo.tal bar.tal" will update foo's and bar's TAs
and delete any other TAs. This is useful when TALs are listed in a
configuration file, to allow updateTA.py to automatically update the
database from the config file even then TALs are removed from the
configuration.

It also has the side effect that if you change the location within a
TAL file and run "updateTA.py -d", the TA cached for the old location
is removed from the database.