it's probably not worth fixing yet
The Mac version of patch(1) fails to apply some of our patches that
are used to generate the conformance test cases. Since it's a failure
to generate test cases, not a failure of our software to run the test
cases, I don't think it's worth fixing for now. It should, however, be
fixed if we later decide to fully support Mac.