move function documentation to function declaration
Move comments documenting functions from the function definitions (.c
files) to the function declarations (.h files).  This convention means
that any installed headers (which describe the public API) will
include API documentaiton.  (The .c files are never installed, so
users won't see the documentation of public APIs if the comments are
in the .c files, unless the comments are converted to Doxygen and the
Doxygen documentation is compiled and installed.)

This change is not comprehensive; there are still some functions that
are documented near the definition.