define xsnprintf() and xvsnprintf() for improved safety
Define functions that wrap around the standard snprintf() and
vsnprintf() functions to check the return value.  If the return value
is negative or greater than or equal to the buffer length call
abort(), otherwise return it as-is.  This makes it safe to blindly use
the return value to increment an index when building up a string.