avoid the non-portable nc -z option
Define a shell function called port_open() that tests to see if some
local process is listening on a given TCP port and call that function
instead of running 'nc -z'.  The port_open() function uses nc, but to
avoid compatibility problems with versions of nc that don't support
the -z option it simply redirects output/input to/from /dev/null.

addresses [#1]