improve the schema for prefixes
This commit does two things to improve how prefixes are stored in the
database:
 1. For ROAs, switch from a single, very large cell per ROA to a
    separate table with one row per prefix. This removes an upper
    bound on the number of prefixes per ROA.
 2. For all prefixes, switch from a textual format to a simple binary
    format.

WARNING: This commit only updates the schema, it does not update any
of the code that relies on the schema. As such, this commit breaks a
lot of code. Subsequent commit(s) will update that code to work again.

addresses [#6] and [#7]