use the new schema when querying rpki-rtr data
Use the schema introduced in the previous set of commits when querying
rpki-rtr data from the database. Additionally, change how rpki-rtr
data is queried during Serial and Reset Queries to significantly
improve performance with large rtr_incremental or rtr_full tables.

NOTE: This is one of a series of commits to use the updated schema.
It is the last in the series to update the non-test code, so the code
should be working now, but the tests will still fail.

addresses [#5] and [#7]