fix certificate expiration dates
These certificates expired in the past, causing `make check` to fail.
Now they expire in 2049, which should be a little while off. I chose
not to use a later year (which would require using GeneralizedTime
instead of UTCTime) because the current test infrastructure doesn't
seem to handle GeneralizedTime very well. This should be fixed some
time before 2049.

I did not update tests/conformance/rfc3779/R.raw because it's
currently unused. I did not update any of the dates in var/templates
because all the places that use the templates override the templates'
dates.