avoid compatibility issues with nc
The port_open() shell function removed by this commit fails with nc
implementations that do not shutdown the write side of the socket on
EOF from stdin. This is because the server doesn't know that the
client is done, and therefore the server waits indefinitely.
Meanwhile, the client is also waiting indefinitely. Some versions of
nc have a -q or -z option that would fix this, but other versions of
nc do not have either of these options.

Therefore, this commit re-writes port_open as a python script instead
of a shell function that uses nc. This commit also adds a tests/bin/
directory for executables used only by the test suite.