add tests for scrub_for_print()'s output length
scrub_for_print() is underdocumented, so it's not clear what
dst_len_out is intended to hold.  However, I'm assuming it's meant to
hold the untruncated length of the escaped string based on the
following evidence:
  * the note "Output might be truncated, compared to input."
  * this function is based on snprintf(), which returns the
    untruncated string length
  * you can already get the actual output length by doing strlen() on
    the destination buffer, so returning the actual length isn't as
    useful (though it helps avoid an O(n) character walk)

These tests currently fail due to bugs in scrub_for_print().
Subsequent commits will fix the bugs.