fix off-by-one buffer truncation bugs
snprintf() truncation occurs if the return value is greater than or
equal to the size of the buffer.  Both write_casn_double() and
roaGenerateFilter2() were only checking to see if the return value was
greater than the buffer length.  Switch to a greater-than-or-equal-to
check to ensure that any truncation is detected.