use strncpy() instead of snprintf() to copy part of a string
Use strncpy() with a carefully chosen length rather than snprintf()
into a carefully undersized buffer.  This should be mildly faster (no
need to process the "%s", and no need to walk the entire source string
to calculate the return value), but the primary reason for the switch
is to improve readability:  snprintf() is almost always used on a
buffer that is expected to be big enough to hold the entire formatted
string, so purposeful truncation with snprintf() is highly unusual.