configure.ac: restore call to AM_PROG_CC_C_O
The call to AM_PROG_CC_C_O was removed in
cbfb0d32ad1aa53d837971563f0aec0f485c4532, but it appears to be
required by older versions of Automake that we still support (versions
older than v1.14, which was released 2013-06-21).

Without it, Automake prints the following:

    mk/libcasn.mk:40: compiling `lib/casn/asn.c' in subdir requires `AM_PROG_CC_C_O' in `configure.ac'
    Makefile.am:80:   `mk/libcasn.mk' included from here

The above doesn't seem to be an error; autoreconf does not exit non-0,
and rpstir still seems to build OK and pass all tests.  But the
message makes me uncomfortable, and it doesn't hurt to add it back.