instruct users to build shared OpenSSL libraries
Building and linking against shared libraries works around a problem
on NetBSD 7.0 where it uses the non-RFC3779 functions from the
system's libcrypt.so instead of the RFC3779-enabled functions from the
statically linked libcrypt.a.

Linux systems don't have this same problem for some reason.  It does
not hurt to build the shared libraries on Linux systems, however.