use an enum for error codes
Enums automatically increment values, which makes it harder to
accidentally assign the same value to two different error codes.
Macros are used to avoid repeating the constants, short names, and
descriptions, which further reduces developer errors.  The new enum
type is named err_code, and this name can be used as a function return
value type or out-parameter type to improve code readability.

One drawback is that it's harder to delete an entry or reorder the
lines without causing entries to get different values.  As long as we
don't depend on specific numbers anywhere (e.g., print them so that
they show up in bug reports) then this shouldn't matter.

Note:  This new enum shifts the values after ERR_SCM_INVALROA by 1
because there was a gap in the #defines.