add a generic "Unspecified error" error code
and use it instead of a hard-coded -1 for places without a specific
error code.  Due to the lack of clear documentation specifying these
functions' return values, it's unclear whether the original authors
intended to return `ERR_SCM_*` values.  However, after auditing all of
the callers, returning `ERR_SCM_UNSPECIFIED` is not inconsistent with
how the callers use the return values.