conservatively check X509_CRL_verify() return value
According to the OpenSSL documentation, X509_CRL_verify() returns 1 on
succes, 0 on failure, and -1 on error.  Rather than check for failure
(by checking whether the return value is non-positive), check for
non-success (test whether it's not equal to 1).  This means that a
return value of 2 or greater will be interpreted as a failure rather
than success, which should be safer given that X509_CRL_verify()
should never return 2 or greater.