fix returning casn function return value as an ERR_SCM_* value
The negative return values from the casn functions are not ERR_SCM_*
values, but callers of the rpki functions expect to get an ERR_SCM_*
value.  Return ERR_SCM_UNSPECIFIED instead of the casn function's
return value.

ERR_SCM_UNSPECIFIED is used because it equals -1, and that's what the
casn functions return on error.  Also add @bug comments to replace
that error code with something more appropriate.