fix NULL dereference in error condition check
If cert_answersp is NULL then it's not valid to dereference the
pointer.  I suspect the original intention was to use '||' instead of
'&&'.  However, rather than change the '&&' to '||', just delete that
extra clause -- there is already a check for less than zero a couple
of lines later, and the return value in that check is more appropriate
(struct cert_answers::num_ansrs is documented to hold an ERR_SCM_*
error code if negative, so it's better to return that value than to
return ERR_SCM_UNSPECIFIED).