stop setting RPKIAllowStaleValidationChain in tests
This option will go away soon, so stop setting it.  This makes some of
the test cases and config files redundant with others, so delete those
redundant tests and config files.

With the deleted reduntant tests, some of the step scripts are now
empty.  They will be deleted and the remaining scripts renumbered in a
separate commit.