deprecate RPKIAllowStaleValidationChain
This option significantly increases implementation complexity without
much apparent value.  The feature depends on the SCM_FLAG_NOCHAIN
flag, which is somewhat meaningless in a world where an evil twin can
create a chain to a TA.  The SCM_FLAG_NOCHAIN flag will go away in a
future commit as a step toward fixing the evil twin bug, so simply
deprecate the RPKIAllowStaleValidationChain option rather than figure
out how to reimplement it without that flag.