use _Bool instead of bool
This header doesn't include stdbool.h or anything that directly or
indirectly includes stdbool.h, so it is not valid to use the bool
typedef.  Fix this by changing bool to _Bool.  (An alternative would
be to include stdbool.h, but that could result in name collisions in
users of this header.)