mark a passing conformance test as expected to pass
The apparent intention of this testcase (based on its name and the
documentation in doc/conformance-cases) is to ensure that a GBR with a
non-vCard payload is rejected.  RPSTIR has never thoroughly validated
GBR payload for vCard conformance, which is why this testcase was
marked as "ignored" ever since it was created.  However, the payload
in this testcase isn't just invalid vCard data, it's also invalid
UTF-8 -- and RPSTIR *does* verify that the payload is valid UTF-8 (but
see note below).  So really this testcase should be named something
like NotVCardBecauseNotUTF8, and we should add additional
NotVCardBecause* testcases.

Note that due to the way the UTF-8 check is implemented, this test
will fail on systems that do not have a C.UTF-8 locale that behaves as
expected.  Most modern systems of interest have such a locale, so no
attempt is made to mark the test as expected to fail on systems
without that locale.  In the future, the implementation of the UTF-8
check should be replaced with something more portable and
standards-conformant.