fix getting pointer to pointer to start of array
According to the C standard, if foo is an array type, foo can also be
used as a pointer to the first element in the array in many contexts.
However, unlike an actual pointer, &foo does not get the address of
the address of the first element; it is the same as foo without the &.

This fixes a bug introduced in b9d846a.