don't run valgrind on interpreted scripts of any kind
If a test program is an interpreted script, the test program should
not be run in valgrind.  Running an interpreter in valgrind is
unlikely to be useful, will unnecessarily slow down the tests, and
will produce lots of uninteresting log output.

Before, this test driver deemed a test program to be an interpreted
script if the filename matched the pattern *.sh.  This heuristic is
flawed:  it is incorrect when a script's filename does not end in .sh
(such as the recently created *.tap scripts) and it is incorrect if a
binary executable's filename ends in .sh (unlikely, but still).

Now this test driver determines if a test program is an interpreted
script by examining the first two bytes of the program file.  If they
are "#!" (shebang) then the test program is deemed to be an
interpreted script.  This magic number test is the same test used by
the loader, so it should be reliable.