skip badGBRNotVCard.gbr conformance test
check_ghostbusters_content() in lib/rpki/cms/roa_validate.c checks to
see if the Ghostbusters payload is valid UTF-8 by temporarily
switching the locale character type to C.UTF-8.  This locale is
non-standard and does not exist on all systems, so RPSTIR skips the
check if that locale is not available.  This causes the
badGBRNotVCard.gbr conformance test case to fail on such systems.

The testcase is non-comprehensive anyway, so simply skip it.